# openapi.yaml
openapi: 3.0.3
info:
  title: Huellitas Solidarias API
  description: >
    API REST para la plataforma de adopción responsable de mascotas,
    fundaciones, solicitudes de adopción, favoritos, mensajes y donaciones.
    Las rutas no llevan prefijo de versión (`/mascotas`, no `/api/v1/mascotas`),
    reflejando el código real del backend.
  version: "1.0.0"
servers:
  - url: http://localhost:3000
    description: Desarrollo local
security:
  - bearerAuth: []
...
paths:
...
  /mascotas:
    get:
      tags: [Mascotas]
      summary: Listado autenticado (fundación ve las propias, resto ve el catálogo visible)
      responses:
        "200": { description: OK }
    post:
      tags: [Mascotas]
      summary: Crear mascota (fundación, admin)
      responses:
        "201": { description: Creada }
        "403": { description: Rol insuficiente }
...
  /solicitudes:
    get:
      tags: [Solicitudes]
      summary: Listar solicitudes de adopción (según rol)
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
        - name: estado
          in: query
          schema: { type: string, enum: [revision, aprobada, rechazada, seguimiento] }
      responses:
        "200": { description: OK }
    post:
      tags: [Solicitudes]
      summary: Crear solicitud de adopción (rol usuario)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [petId, form]
              properties:
                petId: { type: integer }
                form: { type: object }
      responses:
        "201": { description: Creada }
        "400": { description: Mascota no disponible / solicitud duplicada }
...
  /fundaciones:
    get:
      tags: [Fundaciones]
      summary: Listar solicitudes de registro de fundación (admin)
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
        - name: estado
          in: query
          schema: { type: string, enum: [pendiente, aprobada, rechazada] }
      responses:
        "200": { description: OK }
    post:
      tags: [Fundaciones]
      summary: Registrar una solicitud de fundación (público)
      security: []
      responses:
        "201": { description: Creada }
        "409": { description: Correo/RUC duplicados }
...
  /mensajes:
    get:
      tags: [Mensajes]
      summary: Listar mensajes de contacto (fundacion, admin)
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
        - name: leido
          in: query
          schema: { type: boolean }
      responses:
        "200": { description: OK }
        "403": { description: Rol insuficiente }
    post:
      tags: [Mensajes]
      summary: Enviar un mensaje de contacto (público)
      security: []
      responses:
        "201": { description: Creado }
...
  /donaciones:
    get:
      tags: [Donaciones]
      summary: Listar donaciones (admin ve todas, resto las propias)
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
        - name: tipo
          in: query
          schema: { type: string }
        - name: estado
          in: query
          schema: { type: string, enum: [Completado, Pendiente, Cancelado] }
      responses:
        "200": { description: OK }
    post:
      tags: [Donaciones]
      summary: Registrar una donación (público o autenticado)
      security: []
      responses:
        "201": { description: Creada }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
...